alteração de int para float, a fim de aumentar a capacidade de cálculo

+++ funcoes.java
@@ -1,24 +1,24 @@
 public class funcoes{
-    public static int soma(int a, int b){
+    public static float soma(float a, float b){
         return a + b;
     }
 
-    public static int subtracao(int a, int b){
+    public static float subtracao(float a, float b){
         return a - b;
     }
 
-    public static int multiplicacao(int a, int b){
+    public static float multiplicacao(float a, float b){
         return a * b;
     }
 
-    public static int divisao (int a, int b){
+    public static float divisao (float a, float b){
         if (b == 0){
             System.out.println("ERROR: Dividiu por 0");
         }else{
             return a / b;
         }
     }
-    public static int resto (int a, int b){
+    public static float resto (float a, float b){
         return a % b;
     }
